laravel-project

+++ app/Providers/ViewComposers/ProductsComposer.php
@@ -5,9 +5,10 @@ use App\Models\Product;
 class ProductsComposer{
     public function compose(View $view){
         $all = Product::all();
-        $min_price = (float)$all->sortBy('price')->first()->price;
-        $max_price = (float)$all->sortByDesc('price')->first()->price; //sort vice versa
-        $avg_price = ((float)$max_price + (float)$min_price) / 2;
+        $min_price = (int)$all->sortBy('price')->first()->price;
+        $max_price = (int)$all->sortByDesc('price')->first()->price; //sort vice versa
+        // dd($max_price);
+        $avg_price = ((int)$max_price + (float)$min_price) / 2;
         $view->with('min_price', $min_price)->with('max_price', $max_price)->with('avg_price',$avg_price);
     }
 }
